📝 feat. 로그아웃api 수정, 타입 파일 수정

+++ src/containers/header/Header.tsx
@@ -2,16 +2,17 @@ import * as t from "./header.style";
 import { useEffect, useState } from "react";
 import { Link } from "react-router-dom";
 import { SearchBar } from "../../elements/searchBar/SearchBar";
-import headLogo from "./../../assets/logo/headLogo.png";
-import SchProdModal from "../../components/modal/schProdModal/SchProdModal";
-import ScrollHeader from "../../components/header/scroll/ScrollHeader";
-import MenuDrop from "../../elements/menuDrop/MenuDrop";
-import useDropDown from "../../hooks/useDropDown";
-import AsideHeader from "../../components/header/aside/AsideHeader";
+import { delCartDB } from "../../shared/utils/delCartDB";
+import { getAllCartDB } from "../../hooks/useAllCartDB";
 import { useLogOut } from "./useLogout";
 import { useViewport } from "../../hooks/useViewport";
-import { getAllCartDB } from "../../hooks/useAllCartDB";
-import { CartType } from "../../shared/types/types";
+import { CartListType, CartPayType } from "../../shared/types/types";
+import useDropDown from "../../hooks/useDropDown";
+import headLogo from "./../../assets/logo/headLogo.png";
+import MenuDrop from "../../elements/menuDrop/MenuDrop";
+import ScrollHeader from "../../components/header/scroll/ScrollHeader";
+import AsideHeader from "../../components/header/aside/AsideHeader";
+import SchProdModal from "../../components/modal/schProdModal/SchProdModal";
 
 let cateData: {
   id: number;
@@ -35,7 +36,9 @@ const Header = () => {
   const { isDropped, dropRef, handleRemove } = useDropDown();
   const [searchIsOpen, setSearchIsOpen] = useState<boolean>(false);
   const [infoIsOpen, setInfoIsOpen] = useState<boolean>(false);
-  const [cartData, setCartData] = useState<CartType[]>();
+  const [cartData, setCartData] = useState<CartPayType[]>();
+  const [cartList, setCartList] = useState<CartListType[]>();
+  const [idList, setIdList] = useState([]);
 
   const goHome = () => {
     window.location.href = "/";
@@ -43,26 +46,48 @@ const Header = () => {
 
   useEffect(() => {
     const result = getAllCartDB();
-    result.then((res) => setCartData(res));
+    result.then((res) => {
+      setCartData(res);
+    });
   }, []);
 
-  // const CData = {
-  //   cartList: cartData,
-  // };
+  useEffect(() => {
+    if (cartData) {
+      const list = cartData.map((item) => ({
+        k_No: item.keywordNo,
+        p_No: item.productNo,
+        p_Thumbnail: item.thumbnail,
+        p_Category: item.category,
+        a_Brand: item.brand,
+        p_Name: item.name,
+        p_Cost: item.price,
+        p_Discount: item.discount,
+        p_Option: item.option,
+        p_Price: item.totalPrice,
+        p_Cnt: item.totalQty,
+      }));
+      const idList = list.map((item) => item.p_No);
+      setCartList(list);
+      setIdList(idList);
+    }
+  }, [cartData]);
 
-  // const { mutate, isSuccess, data } = useLogOut(CData);
+  const logOutData = {
+    cartList: cartList,
+  };
 
-  // useEffect(() => {
-  //   if (isSuccess) {
-  //     console.log("Success");
-  //     localStorage.clear();
-  //     window.location.href = "/";
-  //   }
-  // }, [isSuccess]);
+  const { mutate, isSuccess } = useLogOut(logOutData);
+
+  useEffect(() => {
+    if (isSuccess) {
+      delCartDB(idList);
+      localStorage.clear();
+      window.location.href = "/";
+    }
+  }, [isSuccess]);
 
   const handelLogOut = () => {
-    localStorage.clear();
-    //   mutate();
+    mutate();
   };
 
   return (
